Introduce FloatPanelModule
* move floatingpanel widget to module
* introduce new event - app\modules\floatPannel\events\BeforeRenderEvent
* move widget hardcode to attachEventsHandlers methods in corresponding
  modules
