+++ application/modules/floatPanel/FloatPanelModule.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace app\modules\floatPanel;
+
+
+use app\components\BaseModule;
+
+class FloatPanelModule extends BaseModule
+{
+
+}
