+++ application/modules/page/PageModule.php
@@ -3,13 +3,17 @@
 namespace app\modules\page;
 
 use app;
+use app\modules\event\interfaces\EventInterface;
+use app\modules\floatPanel\widgets\FloatingPanel;
+use kartik\icons\Icon;
 use Yii;
+use yii\base\Event;
 
 /**
  * Base configuration module for DotPlant2 CMS
  * @package app\modules\page
  */
-class PageModule extends app\components\BaseModule
+class PageModule extends app\components\BaseModule implements EventInterface
 {
     const BACKEND_PAGE_GRID = 'pageGrid';
     /**
@@ -60,5 +64,30 @@ class PageModule extends app\components\BaseModule
                 'label' => Yii::t('app', 'Page edit'),
             ],
         ];
+    }
+
+    /**
+     * @return void
+     */
+    public static function attachEventsHandlers()
+    {
+        Event::on(
+            FloatingPanel::class,
+            FloatingPanel::EVENT_BEFORE_RENDER,
+            function($event) {
+                if (in_array(\Yii::$app->requestedRoute, ['/page/page/show', '/page/page/list']))
+                    if (isset($_GET['id'])) {
+                        $page = app\modules\page\models\Page::findById($_GET['id']);
+                        $event->items[] = [
+                            'label' => Icon::show('pencil') . ' ' . Yii::t('app', 'Edit page'),
+                            'url' => [
+                                '/page/backend/edit',
+                                'id' => $page->id,
+                                'parent_id' =>$page->parent_id,
+                            ],
+                        ];
+                    }
+                }
+            );
     }
 }
