+++ application/modules/shop/ShopModule.php
@@ -7,6 +7,7 @@ use app\backend\BackendModule;
 use app\components\BaseModule;
 use app\modules\shop\handlers\UserHandler;
 use app\modules\shop\models\ConfigConfigurationModel;
+use kartik\icons\Icon;
 use Yii;
 use yii\base\Application;
 use yii\base\BootstrapInterface;
@@ -17,7 +18,7 @@ use yii\web\User;
  * Shop module is the base core module of DotPlant2 CMS handling all common e-commerce features
  * @package app\modules\shop
  */
-class ShopModule extends BaseModule implements BootstrapInterface
+class ShopModule extends BaseModule implements BootstrapInterface, app\modules\event\interfaces\EventInterface
 {
     const BACKEND_PRODUCT_GRID = 'productEditGrid';
     const BACKEND_CATEGORY_GRID = 'categoryEditGrid';
@@ -215,5 +216,80 @@ class ShopModule extends BaseModule implements BootstrapInterface
                 'label' => Yii::t('app', 'Category edit'),
             ],
         ];
+    }
+
+    /**
+     * @return void
+     */
+    public static function attachEventsHandlers()
+    {
+        Event::on(
+            app\modules\floatPanel\widgets\FloatingPanel::class,
+            app\modules\floatPanel\widgets\FloatingPanel::EVENT_BEFORE_RENDER,
+            function ($event) {
+                switch (Yii::$app->requestedRoute) {
+                    case 'shop/product/list':
+                        if (isset($_GET['properties'])) {
+                            $apply_if_params = [];
+                            foreach ($_GET['properties'] as $property_id => $values) {
+                                if (isset($values[0])) {
+                                    $apply_if_params[$property_id] = $values[0];
+                                }
+                            }
+                            if (Yii::$app->response->dynamic_content_trait === true) {
+                                $event->items[] = [
+                                    'label' => Icon::show('puzzle') . ' ' . Yii::t('app', 'Edit Dynamic Content'),
+                                    'url' => [
+                                        '/backend/dynamic-content/edit',
+                                        'id' => Yii::$app->response->matched_dynamic_content_trait_model->id,
+                                    ],
+                                ];
+                            } else {
+                                if (isset($_GET['properties'], $_GET['last_category_id'])) {
+                                    $event->items[] = [
+                                        'label' => Icon::show('puzzle') . ' ' . Yii::t('app', 'Add Dynamic Content'),
+                                        'url' => [
+                                            '/backend/dynamic-content/edit',
+                                            'DynamicContent' => [
+                                                'apply_if_params' => Json::encode($apply_if_params),
+                                                'apply_if_last_category_id' => $_GET['last_category_id'],
+                                                'object_id' => Object::getForClass(app\modules\shop\models\Product::className())->id,
+                                                'route' => 'shop/product/list',
+                                            ]
+                                        ],
+                                    ];
+
+                                }
+
+                            }
+                        } else {
+                            // no properties selected - go to category edit page
+                            if (isset($_GET['last_category_id'])) {
+                                $cat = app\modules\shop\models\Category::findById($_GET['last_category_id']);
+                                $event->items[] = [
+                                    'label' => Icon::show('pencil') . ' ' . Yii::t('app', 'Edit category'),
+                                    'url' => [
+                                        '/shop/backend-category/edit',
+                                        'id' => $cat->id,
+                                        'parent_id' => $cat->parent_id,
+                                    ],
+                                ];
+                            }
+                        }
+                        break;
+                    case 'shop/product/show':
+                        if (isset($_GET['model_id'])) {
+                            $event->items[] = [
+                                'label' => Icon::show('pencil') . ' ' . Yii::t('app', 'Edit product'),
+                                'url' => [
+                                    '/shop/backend-product/edit',
+                                    'id' => intval($_GET['model_id'])
+                                ],
+                            ];
+                        }
+                        break;
+                }
+            }
+        );
     }
 }
